move most of the core of populate.py into populate_lib.py
This allows the code in the newly created populate_lib.py to be imported as a file in a module consistently.  Currently, populate.py is importing it as a file in the same dir and tests/test_indexer is importing it as a module through a search path hack.
